rename `resolveShape2DProps.ts` to `mergeShape2DProps.ts`
