+++ src/components/Object2D/Shape2D/Path2D/Line.tsx
@@ -11,7 +11,7 @@ import { createTransformedPath } from 'src/utils/createTransformedPath'
 import { createUpdatedContext } from 'src/utils/createUpdatedContext'
 import hitTest from 'src/utils/hitTest'
 import renderPath from 'src/utils/renderPath'
-import { mergeShape2DProps } from 'src/utils/resolveShape2DProps'
+import { mergeShape2DProps } from 'src/utils/mergeShape2DProps'
 import { createControlledProps } from 'src/utils/createControlledProps'
 /**
  * Paints a straight line to the canvas
