+++ src/components/Object2D/Shape2D/Path2D/Path.tsx
@@ -13,7 +13,7 @@ import { createUpdatedContext } from 'src/utils/createUpdatedContext'
 import hitTest from 'src/utils/hitTest'
 import renderPath from 'src/utils/renderPath'
 import { createControlledProps } from 'src/utils/createControlledProps'
-import { mergeShape2DProps } from 'src/utils/resolveShape2DProps'
+import { mergeShape2DProps } from 'src/utils/mergeShape2DProps'
 
 /**
  * Paints a straight line to the canvas
