+++ src/components/Object2D/Shape2D/Text.tsx
@@ -23,7 +23,7 @@ import { createParenthood } from 'src/utils/createParenthood'
 import { createUpdatedContext } from 'src/utils/createUpdatedContext'
 import { resolveExtendedColor } from 'src/utils/resolveColor'
 import { createControlledProps } from 'src/utils/createControlledProps'
-import { mergeShape2DProps } from 'src/utils/resolveShape2DProps'
+import { mergeShape2DProps } from 'src/utils/mergeShape2DProps'
 import { Normalize } from 'src/utils/typehelpers'
 import { RectangleProps } from './Path2D/Rectangle'
