+++ src/utils/mergeShape2DProps.ts
